+++ MonkeyDelivery/Src/Enemies/Bat.cpp
@@ -67,7 +67,7 @@ void Bat::onPlayerInteraction(Player* player)
 		if (stealTimer_ + 750 < SDL_GetTicks()) {
 			player->removeMoney(1);
 
-			sdlutils().soundEffects().at("bat").setVolume(game->getSoundEfectsVolume() * game->getGeneralVolume());
+			sdlutils().soundEffects().at("bat").setVolume((int)(game->getSoundEfectsVolume() * game->getGeneralVolume()));
 			sdlutils().soundEffects().at("bat").play(0, 1);
 			stealTimer_ = SDL_GetTicks();
 		}
